Add Random Recursive Division Generation Algorithm
Short Summary:
-Pick two random nodes
-Pick random neighbours to add to each node
-Once all nodes in the maze are associated with one of the two nodes, Add a line down the division of the two groups
-Repeat these steps with each group
-When there are 2 or less nodes in an area you can no longer split it

Along with this algorithm, added the ability to set the maze to stop updating the maze one node at a time.
When this flag is set, it will group all following node updates
When the flag is set back, it will update all the node changes at the same time.