Convert draw method from divs to svg
I decided I would begin to learn how svgs work.
Therefore I am converting the draw machanism from a group of divs to an svg.
I also made the start and end points use an svg arrow and fixed an issue where the path of the start item was 0 instead of 1